requirements.txt
# Core data & ML
streamlit>=1.0
pandas>=1.3
numpy>=1.21
scikit-learn>=1.0
statsmodels>=0.13

# Deep learning
keras>=2.8
torch>=2.0
pytorch-lightning>=2.0
pytorch-forecasting>=0.13

# Hugging Face (PatchTST support)
transformers[time-series]>=4.38
datasets>=2.10
